feature:S3C-4673 Moving security code to vault-admin-client module for Vault admin changes
