+++ osis-core/src/main/java/com/scality/osis/utils/ScalityConstants.java
@@ -116,10 +116,4 @@ public final class ScalityConstants {
 
     public static final long DEFAULT_MIN_OFFSET = 0l;
     public static final long DEFAULT_MAX_LIMIT = 1000l;
-
-    // Cipher Constants
-    public static final String NAME_AES_256_GCM_CIPHER = "AES256GCM";
-    public final static int DEFAULT_AES_GCM_NONCE_LENGTH = 12;
-    public final static int DEFAULT_AES_GCM_TAG_LENGTH = 128; //128 bit auth tag length
-    public final static int DEFAULT_AES_GCM_256_KEY_LENGTH = 32; //128 bit auth tag length
 }
