+++ vault-admin-client/src/main/java/com/scality/osis/security/crypto/model/SecretKeyRepoData.java
