+++ vault-admin-client/src/main/java/com/scality/osis/security/utils/SecurityConstants.java
@@ -1,0 +1,10 @@
+package com.scality.osis.security.utils;
+
+public class SecurityConstants {
+
+    // Cipher Constants
+    public static final String NAME_AES_256_GCM_CIPHER = "AES256GCM";
+    public final static int DEFAULT_AES_GCM_NONCE_LENGTH = 12;
+    public final static int DEFAULT_AES_GCM_TAG_LENGTH = 128; //128 bit auth tag length
+    public final static int DEFAULT_AES_GCM_256_KEY_LENGTH = 32; //128 bit auth tag length
+}
