+++ vault-admin-client/src/test/java/com/scality/osis/security/crypto/AES256GCMTest.java
@@ -1,17 +1,13 @@
 package com.scality.osis.security.crypto;
 
 import com.scality.osis.security.crypto.model.SecretKeyRepoData;
+import com.scality.osis.security.utils.ScalitySecurityTestUtils;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import javax.crypto.AEADBadTagException;
 import java.security.InvalidKeyException;
-
-import static com.scality.osis.utils.ScalityTestUtils.TEST_ACCESS_KEY;
-import static com.scality.osis.utils.ScalityTestUtils.TEST_CIPHER_SECRET_KEY;
-import static com.scality.osis.utils.ScalityTestUtils.TEST_INVALID_CIPHER_SECRET_KEY;
-import static com.scality.osis.utils.ScalityTestUtils.TEST_SECRET_KEY;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -27,35 +23,35 @@ public class AES256GCMTest {
     @Test
     public void testEncryptDecrypt() throws Exception {
         //create new random key
-        final String key = TEST_CIPHER_SECRET_KEY;
+        final String key = ScalitySecurityTestUtils.TEST_CIPHER_SECRET_KEY;
         
 
-        final SecretKeyRepoData encryptedRepoData = aes256GCMUnderTest.encrypt(TEST_SECRET_KEY, key, TEST_ACCESS_KEY);
+        final SecretKeyRepoData encryptedRepoData = aes256GCMUnderTest.encrypt(ScalitySecurityTestUtils.TEST_SECRET_KEY, key, ScalitySecurityTestUtils.TEST_ACCESS_KEY);
         assertNotNull(encryptedRepoData);
 
-        final String decrypted = aes256GCMUnderTest.decrypt(encryptedRepoData, key, TEST_ACCESS_KEY);
-        assertEquals(TEST_SECRET_KEY, decrypted);
+        final String decrypted = aes256GCMUnderTest.decrypt(encryptedRepoData, key, ScalitySecurityTestUtils.TEST_ACCESS_KEY);
+        Assertions.assertEquals(ScalitySecurityTestUtils.TEST_SECRET_KEY, decrypted);
     }
 
     @Test
     public void testEncryptThrowsException() {
         // Setup
         //create new random key of size 40
-        final String key = TEST_INVALID_CIPHER_SECRET_KEY;
+        final String key = ScalitySecurityTestUtils.TEST_INVALID_CIPHER_SECRET_KEY;
 
 
         // Run the test
-        assertThrows(InvalidKeyException.class, () -> aes256GCMUnderTest.encrypt(TEST_SECRET_KEY, key, TEST_ACCESS_KEY));
+        assertThrows(InvalidKeyException.class, () -> aes256GCMUnderTest.encrypt(ScalitySecurityTestUtils.TEST_SECRET_KEY, key, ScalitySecurityTestUtils.TEST_ACCESS_KEY));
     }
 
     @Test
     public void testDecryptInvalidAssociatedDataThrowsException() throws Exception {
         // Setup
         //create new random key
-        final String key = TEST_CIPHER_SECRET_KEY;
+        final String key = ScalitySecurityTestUtils.TEST_CIPHER_SECRET_KEY;
         
 
-        final SecretKeyRepoData encryptedRepoData = aes256GCMUnderTest.encrypt(TEST_SECRET_KEY, key, TEST_ACCESS_KEY);
+        final SecretKeyRepoData encryptedRepoData = aes256GCMUnderTest.encrypt(ScalitySecurityTestUtils.TEST_SECRET_KEY, key, ScalitySecurityTestUtils.TEST_ACCESS_KEY);
         // Run the test
         assertThrows(AEADBadTagException.class, () -> aes256GCMUnderTest.decrypt(encryptedRepoData, key, "InvalidData"));
     }
