+++ vault-admin-client/src/test/java/com/scality/osis/security/utils/ScalitySecurityTestUtils.java
@@ -1,0 +1,16 @@
+package com.scality.osis.security.utils;
+
+public final class ScalitySecurityTestUtils {
+
+    public static final String TEST_ACCESS_KEY = "access_key";
+    public static final String TEST_SECRET_KEY = "secret_key";
+
+    public static final String TEST_CIPHER_SECRET_KEY = "dGhpc2lzYXJlYWxseWxvbmdhbmRzZXk=";
+    public static final String TEST_INVALID_CIPHER_SECRET_KEY = "dGhpc2lzYXJlYWxseWxvbmdhbmRzZXk=hgJfad==";
+    public static final String TEST_CIPHER_ID = "1";
+
+
+    private ScalitySecurityTestUtils(){
+
+    }
+}
